Deep-Linking support for component search
The query for the component search can be shared and linked to via the
URL (e.g.
`components#/search/COORDINATES/group=acme/name=test/version=dev`).

Performing a search by clicking on the `search` button changes the URL
in the browser address bar to easily share the current search criteria.

Signed-off-by: RBickert <rbt@mm-software.com>