Add Neccesary parameter to resolveVulnAliases function
common.resolveVulnALiases needs the vulnerability source and the aliases to deduplicate the list on the Aliases column,

In case before, the only parameter added to the function was "value" which do not include neither the source of the current
vulnerability or their aliases.

Thats why I am adding row.source and row.aliases to the function.

I discovered that the 'return' statement was located within the condition if(typeof value !== 'undefined'),
This created a problem because the function would not return anything for null values.
To fix this, I moved the 'return' outside of the condition.

Signed-off-by: Andres Tito <andres.tito@rohde-schwarz.com>