Fix redirect loop when authenticated user is lacking permissions
After successful login, users are redirected to the dashboard. The dashboard requires the `VIEW_PORTFOLIO` permission, and redirects back to the login page if that permission is not present. This could lead to redirect loops for users who lack the permission.

Display a popup instead that informs the user about the lack of permissions to proceed.

Relates to https://github.com/DependencyTrack/dependency-track/issues/5098
Backports https://github.com/DependencyTrack/hyades-frontend/pull/308

Co-authored-by: Sahiba Mittal <sahiba.mittal@citi.com>
Signed-off-by: nscuro <nscuro@protonmail.com>